move MetalMaterials to its own file
