+++ slack_data/utilities/materials.py
@@ -1,0 +1,9 @@
+from enum import Enum
+
+
+class MetalMaterial(str, Enum):
+    ALUMINUM = "Aluminum"
+    STEEL = "Steel"
+    STAINLESS_STEEL = "Stainless Steel"
+    TITANIUM = "Titanium"
+    OTHER = "Other"
